TASK-50322 Improve JSON Response writing to not block HTTP connector (#22)
Prior to this change, the JSON response transforming process writes values directly on HTTP Connector Output Stream, which is a thread safe and blocker operation. Thus, this fix will improve writing process to use a buffered writer with a default buffer size (1MB) that can be customized using exo.properties.